requirements.txt
mcp
httpx
python-dotenv
openai
pandas
groq
